fix(desktop): keep the merge-detection test off the node-pty graph
